feat(deep_causality_quantum): new crate — quantum-information kernels move out of physics (Phase 1)
BREAKING (deep_causality_physics): the quantum-information layer moves to the
new workspace crate deep_causality_quantum: QuantumGates / QuantumOps<R>, the
Haruna logical gates (arXiv:2511.15224), the born_probability /
expectation_value / apply_gate / commutator / fidelity kernels with their
PropagatingEffect wrappers, and the Operator/Gate aliases. klein_gordon (a
relativistic field-theory PDE kernel) STAYS in physics; kernels/quantum splits
accordingly.

- Errors: moved kernels return the crate-local QuantumError (outer newtype
  over QuantumErrorEnum, typed variants incl. CommutatorNonZero naming the
  offending node pair); From<QuantumError> for CausalityError bridges the monad.
- born_probability/fidelity wrappers now carry the clamped plain R on the
  value channel (physics' Probability<R> quantity type stays in physics).
- MaybeParallel bound dropped from the moved kernels (multivector needs no
  deep_causality_par; the marker was a physics-side convention).
- HilbertState stays in deep_causality_multivector; metric signatures come
  from deep_causality_metric (SSOT).
- Call sites re-pointed: ikkt_matrix_model, multi_physics_pipeline.
