+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
     "deep_causality_num_dual",
     "deep_causality_par",
     "deep_causality_physics",
+    "deep_causality_quantum",
     "deep_causality_rand",
     "deep_causality_sparse",
     "deep_causality_tensor",
