+++ deep_causality_physics/src/kernels/quantum/mod.rs
@@ -3,12 +3,13 @@
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 
-pub(crate) mod gates;
-pub(crate) mod gates_haruna;
+// The quantum-information kernels (gates, Haruna gates, Dirac-notation ops and
+// their PropagatingEffect wrappers) live in `deep_causality_quantum`. What
+// remains here is the Klein-Gordon PDE kernel — a relativistic field-theory
+// operator on a simplicial manifold, kept with the other physics kernels.
+
 pub(crate) mod mechanics;
 pub(crate) mod wrappers;
 
-pub use gates::*;
-pub use gates_haruna::*;
 pub use mechanics::*;
 pub use wrappers::*;
