+++ deep_causality_physics/tests/BUILD.bazel
@@ -304,9 +304,6 @@ rust_test_suite(
         # Crate to test
         "//deep_causality_physics",
         # Internal dependencies
-        "//deep_causality_multivector",
-        "//deep_causality_num",
-        "//deep_causality_num_complex",
         "//deep_causality_tensor",
         "//deep_causality_topology",
     ],
