+++ deep_causality_quantum/BUILD.bazel
@@ -1,0 +1,36 @@
+load("@rules_rust//rust:defs.bzl", "rust_doc", "rust_doc_test", "rust_library")
+
+rust_library(
+    name = "deep_causality_quantum",
+    srcs = glob([
+        "src/**",
+    ]),
+    crate_root = "src/lib.rs",
+    tags = [
+        "quantum",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//deep_causality_algebra",
+        "//deep_causality_core",
+        "//deep_causality_haft",
+        "//deep_causality_metric",
+        "//deep_causality_multivector",
+        "//deep_causality_num",
+        "//deep_causality_num_complex",
+    ],
+)
+
+rust_doc(
+    name = "doc",
+    crate = ":deep_causality_quantum",
+    tags = ["doc"],
+    visibility = ["//visibility:public"],
+)
+
+rust_doc_test(
+    name = "doc_test",
+    crate = ":deep_causality_quantum",
+    tags = ["doc-test"],
+    visibility = ["//visibility:public"],
+)
