+++ deep_causality_quantum/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "deep_causality_quantum"
+version = "0.1.0"
+edition = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+repository = "https://github.com/deepcausality/deep_causality.rs"
+authors = ["Marvin Hansen <marvin.hansen@gmail.com>"]
+description = "Quantum causal models on the causal monad: quantum-information kernels, gates, and the operator layer for DeepCausality."
+documentation = "https://docs.rs/deep_causality"
+categories = ["science", "simulation"]
+keywords = ["quantum", "causality", "simulation", "science"]
+exclude = [
+    "*.bazel",
+    "*/*.bazel",
+    "*.bazel.*",
+    "BUILD",
+    "BUILD.bazel",
+    "MODULE.bazel",
+    ".bazelignore",
+    ".bazelrc",
+    "tests/**/*",
+    "papers/*",
+]
+
+[dependencies.deep_causality_algebra]
+path = "../deep_causality_algebra"
+version = "0.1"
+
+[dependencies.deep_causality_core]
+path = "../deep_causality_core"
+version = "0.11.0"
+
+[dependencies.deep_causality_haft]
+path = "../deep_causality_haft"
+version = "0.4"
+
+[dependencies.deep_causality_metric]
+path = "../deep_causality_metric"
+version = "0.2"
+
+[dependencies.deep_causality_multivector]
+path = "../deep_causality_multivector"
+version = "0.5"
+
+[dependencies.deep_causality_num]
+path = "../deep_causality_num"
+version = "0.4"
+
+[dependencies.deep_causality_num_complex]
+path = "../deep_causality_num_complex"
+version = "0.1"
+
+[lints]
+workspace = true
