+++ deep_causality_quantum/tests/error/mod.rs
@@ -3,6 +3,4 @@
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 #[cfg(test)]
-mod mechanics_tests;
-#[cfg(test)]
-mod wrappers_tests;
+mod quantum_error_tests;
