+++ deep_causality_quantum/tests/kernels/mod.rs
@@ -3,6 +3,10 @@
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 #[cfg(test)]
+mod gates_haruna_tests;
+#[cfg(test)]
+mod gates_tests;
+#[cfg(test)]
 mod mechanics_tests;
 #[cfg(test)]
 mod wrappers_tests;
