+++ deep_causality_quantum/tests/mod.rs
@@ -2,7 +2,5 @@
  * SPDX-License-Identifier: MIT
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
-#[cfg(test)]
-mod mechanics_tests;
-#[cfg(test)]
-mod wrappers_tests;
+mod error;
+mod kernels;
