+++ examples/quantum_examples/ikkt_matrix_model/main.rs
@@ -20,7 +20,7 @@
 use deep_causality_algebra::DivisionAlgebra;
 use deep_causality_multivector::{HilbertState, Metric};
 use deep_causality_num_complex::Complex;
-use deep_causality_physics::{Operator, commutator_kernel};
+use deep_causality_quantum::{Operator, commutator_kernel};
 
 /// Switch this alias to `f32` for low precision, `f64` for standard precision,
 /// or `Float106` for high precision.
